gpu/requirements.txt
# GPU Price Maximizer - Dependencies

# Oracle (Agentic - Web Search + GitHub Models)
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Legacy (Playwright - kept for reference)
playwright>=1.40.0
playwright-stealth>=1.0.0
